Extract seed editing hook
